fix: refactor for comments
