+++ halo2_proofs/src/plonk/prover.rs
@@ -5,322 +5,27 @@ use group::{prime::PrimeCurveAffine, Curve};
 use rand_core::RngCore;
 
 use std::hash::Hash;
-use std::marker::PhantomData;
-use std::ops::RangeTo;
+use std::iter;
 
 #[cfg(feature = "multicore")]
 use crate::multicore::IndexedParallelIterator;
 use crate::multicore::{IntoParallelIterator, ParallelIterator};
-use std::{collections::HashMap, iter};
+use std::collections::HashMap;
 
 use super::{
-    circuit::{
-        sealed::{self},
-        Advice, Any, Assignment, Challenge, Circuit, Column, ConstraintSystem, Fixed, FloorPlanner,
-        Instance, Selector,
-    },
-    lookup, permutation, vanishing, ChallengeBeta, ChallengeGamma, ChallengeTheta, ChallengeX,
-    ChallengeY, Error, ProvingKey,
+    circuit::Circuit, lookup, permutation, vanishing, witness, ChallengeBeta, ChallengeGamma,
+    ChallengeTheta, ChallengeX, ChallengeY, Error, ProvingKey,
 };
+use super::witness::{build_instance_single, AdviceCommitted, InstanceSingle};
 
 use crate::{
     arithmetic::{eval_polynomial, CurveAffine},
-    circuit::Value,
-    plonk::Assigned,
     poly::{
         commitment::{Blind, CommitmentScheme, Params, Prover},
-        Basis, Coeff, LagrangeCoeff, Polynomial, ProverQuery,
+        Coeff, LagrangeCoeff, Polynomial, ProverQuery,
     },
 };
-use crate::{
-    poly::batch_invert_assigned,
-    transcript::{EncodedChallenge, TranscriptWrite},
-};
-
-/// Per-circuit instance state produced by phase-1 synthesis.
-///
-/// Mirrors the halo2-gpu fork's `InstanceSingle` shape so downstream crates can
-/// name the same struct across cpu/cuda builds. `instance_polys` is the
-/// coefficient-basis form of `instance_values` (populated eagerly during phase 1).
-#[derive(Debug)]
-pub struct InstanceSingle<C: CurveAffine> {
-    pub instance_values: Vec<Polynomial<C::Scalar, LagrangeCoeff>>,
-    pub instance_polys: Vec<Polynomial<C::Scalar, Coeff>>,
-}
-
-/// Per-circuit advice state produced by [`synthesize_witness`].
-///
-/// Mirrors the halo2-gpu fork's `AdviceSingle` shape (both bases, no blinds).
-/// `advice_values` is the Lagrange-basis witness; `advice_polys` is the same
-/// data in coefficient basis. Blinds are drawn but not retained here — they
-/// are only needed on the internal proving path.
-#[derive(Debug)]
-pub struct AdviceSingle<C: CurveAffine> {
-    pub advice_values: Vec<Polynomial<C::Scalar, LagrangeCoeff>>,
-    pub advice_polys: Vec<Polynomial<C::Scalar, Coeff>>,
-}
-
-#[derive(Clone)]
-struct AdviceCommitted<C: CurveAffine, B: Basis> {
-    pub advice_polys: Vec<Polynomial<C::Scalar, B>>,
-    pub advice_blinds: Vec<Blind<C::Scalar>>,
-}
-
-/// The shape of advice columns accepted by [`create_proof_raw`]: one
-/// `Vec<F>` per physical advice column, each of length `params.n()`.
-///
-/// Aliased so downstream crates can name it unconditionally: the halo2-gpu
-/// fork exposes the same type name at the same path with a device-resident
-/// buffer element type. Callers reading `halo2_proofs::plonk::AdviceColumns`
-/// therefore stay portable across the cpu/cuda forks without a feature gate.
-pub type AdviceColumns<F> = Vec<Vec<F>>;
-
-struct WitnessCollection<'params, 'a, 'b, Scheme, P, C, E, R, T>
-where
-    Scheme: CommitmentScheme<Curve = C>,
-    P: Prover<'params, Scheme>,
-    C: CurveAffine,
-    E: EncodedChallenge<C>,
-    R: RngCore + 'a,
-    T: TranscriptWrite<C, E>,
-{
-    params: &'params Scheme::ParamsProver,
-    current_phase: sealed::Phase,
-    advice: Vec<Polynomial<Assigned<C::Scalar>, LagrangeCoeff>>,
-    challenges: &'b mut HashMap<usize, C::Scalar>,
-    instances: &'b [&'a [C::Scalar]],
-    usable_rows: RangeTo<usize>,
-    advice_single: AdviceCommitted<C, LagrangeCoeff>,
-    instance_single: &'b InstanceSingle<C>,
-    rng: &'b mut R,
-    transcript: &'b mut &'a mut T,
-    column_indices: [Vec<usize>; 3],
-    challenge_indices: [Vec<usize>; 3],
-    unusable_rows_start: usize,
-    _marker: PhantomData<(P, E)>,
-}
-
-impl<'params, 'a, 'b, F, Scheme, P, C, E, R, T> Assignment<F>
-    for WitnessCollection<'params, 'a, 'b, Scheme, P, C, E, R, T>
-where
-    F: Field,
-    Scheme: CommitmentScheme<Curve = C>,
-    P: Prover<'params, Scheme>,
-    C: CurveAffine<ScalarExt = F>,
-    E: EncodedChallenge<C>,
-    R: RngCore,
-    T: TranscriptWrite<C, E>,
-    <Scheme as CommitmentScheme>::ParamsProver: Sync,
-{
-    fn enter_region<NR, N>(&mut self, _: N)
-    where
-        NR: Into<String>,
-        N: FnOnce() -> NR,
-    {
-        // Do nothing; we don't care about regions in this context.
-    }
-
-    fn exit_region(&mut self) {
-        // Do nothing; we don't care about regions in this context.
-    }
-
-    fn enable_selector<A, AR>(&mut self, _: A, _: &Selector, _: usize) -> Result<(), Error>
-    where
-        A: FnOnce() -> AR,
-        AR: Into<String>,
-    {
-        // We only care about advice columns here
-
-        Ok(())
-    }
-
-    fn annotate_column<A, AR>(&mut self, _annotation: A, _column: Column<Any>)
-    where
-        A: FnOnce() -> AR,
-        AR: Into<String>,
-    {
-        // Do nothing
-    }
-
-    fn query_instance(&self, column: Column<Instance>, row: usize) -> Result<Value<F>, Error> {
-        if !self.usable_rows.contains(&row) {
-            return Err(Error::not_enough_rows_available(self.params.k()));
-        }
-
-        self.instances
-            .get(column.index())
-            .and_then(|column| column.get(row))
-            .map(|v| Value::known(*v))
-            .ok_or(Error::BoundsFailure)
-    }
-
-    fn assign_advice<'v>(
-        //<V, VR, A, AR>(
-        &mut self,
-        //_: A,
-        column: Column<Advice>,
-        row: usize,
-        to: Value<Assigned<F>>,
-    ) -> Value<&'v Assigned<F>> {
-        // debug_assert_eq!(self.current_phase, column.column_type().phase);
-
-        debug_assert!(
-            self.usable_rows.contains(&row),
-            "{:?}",
-            Error::not_enough_rows_available(self.params.k())
-        );
-
-        let advice_get_mut = self
-            .advice
-            .get_mut(column.index())
-            .expect("Not enough advice columns")
-            .get_mut(row)
-            .expect("Not enough rows");
-        // We can get another 3-4% decrease in witness gen time by using the following unsafe code, but this skips all array bound checks so we should use it only if the performance gain is really necessary:
-        /*
-        let advice_get_mut = unsafe {
-            self.advice
-                .get_unchecked_mut(column.index())
-                .get_unchecked_mut(row)
-        };
-        */
-        *advice_get_mut = to
-            .assign()
-            .expect("No Value::unknown() in advice column allowed during create_proof");
-        let immutable_raw_ptr = advice_get_mut as *const Assigned<F>;
-        Value::known(unsafe { &*immutable_raw_ptr })
-    }
-
-    fn assign_fixed(&mut self, _: Column<Fixed>, _: usize, _: Assigned<F>) {
-        // We only care about advice columns here
-    }
-
-    fn copy(&mut self, _: Column<Any>, _: usize, _: Column<Any>, _: usize) {
-        // We only care about advice columns here
-    }
-
-    fn fill_from_row(
-        &mut self,
-        _: Column<Fixed>,
-        _: usize,
-        _: Value<Assigned<F>>,
-    ) -> Result<(), Error> {
-        Ok(())
-    }
-
-    fn get_challenge(&self, challenge: Challenge) -> Value<F> {
-        self.challenges
-            .get(&challenge.index())
-            .cloned()
-            .map(Value::known)
-            .unwrap_or_else(Value::unknown)
-    }
-
-    fn push_namespace<NR, N>(&mut self, _: N)
-    where
-        NR: Into<String>,
-        N: FnOnce() -> NR,
-    {
-        // Do nothing; we don't care about namespaces in this context.
-    }
-
-    fn pop_namespace(&mut self, _: Option<String>) {
-        // Do nothing; we don't care about namespaces in this context.
-    }
-
-    fn next_phase(&mut self) {
-        let phase = self.current_phase.to_u8() as usize;
-        #[cfg(feature = "profile")]
-        let start1 = start_timer!(|| format!("Phase {phase} inversion and MSM commitment"));
-        if phase == 0 {
-            // Absorb instances into transcript.
-            // Do this here and not earlier in case we want to be able to mutate
-            // the instances during synthesize in FirstPhase in the future
-            if !P::QUERY_INSTANCE {
-                for values in self.instances.iter() {
-                    for value in values.iter() {
-                        self.transcript
-                            .common_scalar(*value)
-                            .expect("Absorbing instance value to transcript failed");
-                    }
-                }
-            } else {
-                let instance_commitments_projective: Vec<_> =
-                    (&self.instance_single.instance_values)
-                        .into_par_iter()
-                        .map(|poly| self.params.commit_lagrange(poly, Blind::default()))
-                        .collect();
-                let mut instance_commitments =
-                    vec![C::identity(); instance_commitments_projective.len()];
-                C::CurveExt::batch_normalize(
-                    &instance_commitments_projective,
-                    &mut instance_commitments,
-                );
-                let instance_commitments = instance_commitments;
-                drop(instance_commitments_projective);
-
-                for commitment in &instance_commitments {
-                    self.transcript
-                        .common_point(*commitment)
-                        .expect("Absorbing instance commitment to transcript failed");
-                }
-            }
-        }
-        // Commit the advice columns in the current phase
-        let mut advice_values = batch_invert_assigned(
-            self.column_indices
-                .get(phase)
-                .expect("The API only supports 3 phases right now")
-                .iter()
-                .map(|column_index| &self.advice[*column_index][..])
-                .collect(),
-        );
-        // Add blinding factors to advice columns
-        for advice_values in &mut advice_values {
-            for cell in &mut advice_values[self.unusable_rows_start..] {
-                *cell = F::random(&mut self.rng);
-            }
-        }
-        // Compute commitments to advice column polynomials
-        let blinds: Vec<_> = advice_values
-            .iter()
-            .map(|_| Blind(F::random(&mut self.rng)))
-            .collect();
-        let advice_commitments_projective: Vec<_> = (&advice_values)
-            .into_par_iter()
-            .zip((&blinds).into_par_iter())
-            .map(|(poly, blind)| self.params.commit_lagrange(poly, *blind))
-            .collect();
-        let mut advice_commitments = vec![C::identity(); advice_commitments_projective.len()];
-        C::CurveExt::batch_normalize(&advice_commitments_projective, &mut advice_commitments);
-        let advice_commitments = advice_commitments;
-        drop(advice_commitments_projective);
-
-        for commitment in &advice_commitments {
-            self.transcript
-                .write_point(*commitment)
-                .expect("Absorbing advice commitment to transcript failed");
-        }
-        for ((column_index, advice_poly), blind) in self.column_indices[phase]
-            .iter()
-            .zip(advice_values)
-            .zip(blinds)
-        {
-            self.advice_single.advice_polys[*column_index] = advice_poly;
-            self.advice_single.advice_blinds[*column_index] = blind;
-        }
-        for challenge_index in self.challenge_indices[phase].iter() {
-            let existing = self.challenges.insert(
-                *challenge_index,
-                *self.transcript.squeeze_challenge_scalar::<()>(),
-            );
-            assert!(existing.is_none());
-        }
-        self.current_phase = self.current_phase.next();
-        #[cfg(feature = "profile")]
-        end_timer!(start1);
-    }
-}
+use crate::transcript::{EncodedChallenge, TranscriptWrite};
 
 /// This creates a proof for the provided `circuit` when given the public
 /// parameters `params` and the proving key [`ProvingKey`] that was
@@ -341,7 +46,7 @@ pub fn create_proof<
     circuits: &[ConcreteCircuit],
     instances: &[&[&'a [Scheme::Scalar]]],
     mut rng: R,
-    mut transcript: &'a mut T,
+    transcript: &'a mut T,
 ) -> Result<(), Error>
 where
     Scheme::Scalar: Hash + WithSmallOrderMulGroup<3>,
@@ -361,136 +66,24 @@ where
     pk.vk.hash_into(transcript)?;
 
     let domain = &pk.vk.domain;
-    let mut meta = ConstraintSystem::default();
-    #[cfg(feature = "circuit-params")]
-    let config = ConcreteCircuit::configure_with_params(&mut meta, circuits[0].params());
-    #[cfg(not(feature = "circuit-params"))]
-    let config = ConcreteCircuit::configure(&mut meta);
-
-    // Selector optimizations cannot be applied here; use the ConstraintSystem
-    // from the verification key.
     let meta = &pk.vk.cs;
 
     let instance: Vec<InstanceSingle<Scheme::Curve>> = instances
         .iter()
-        .map(|instance| -> InstanceSingle<Scheme::Curve> {
-            let instance_values = instance
-                .iter()
-                .map(|values| {
-                    let mut poly = domain.empty_lagrange();
-                    assert_eq!(poly.len(), params.n() as usize);
-                    if values.len() > (poly.len() - (meta.blinding_factors() + 1)) {
-                        panic!("Error::InstanceTooLarge");
-                    }
-                    for (poly, value) in poly.iter_mut().zip(values.iter()) {
-                        *poly = *value;
-                    }
-                    poly
-                })
-                .collect::<Vec<_>>();
-
-            let instance_polys: Vec<_> = instance_values
-                .iter()
-                .map(|poly| {
-                    let lagrange_vec = domain.lagrange_from_vec(poly.to_vec());
-                    domain.lagrange_to_coeff(lagrange_vec)
-                })
-                .collect();
-
-            InstanceSingle {
-                instance_values,
-                instance_polys,
-            }
-        })
+        .map(|instance| build_instance_single::<Scheme>(params, domain, meta, instance))
         .collect();
-
-    let mut column_indices = [(); 3].map(|_| vec![]);
-    for (index, phase) in meta.advice_column_phase.iter().enumerate() {
-        column_indices[phase.to_u8() as usize].push(index);
-    }
-    let mut challenge_indices = [(); 3].map(|_| vec![]);
-    for (index, phase) in meta.challenge_phase.iter().enumerate() {
-        challenge_indices[phase.to_u8() as usize].push(index);
-    }
 
     #[cfg(feature = "profile")]
     let phase1_time = start_timer!(|| "Phase 1: Witness assignment and MSM commitments");
-    let (advice, challenges) = {
-        let mut advice = Vec::with_capacity(instances.len());
-        let mut challenges = HashMap::<usize, Scheme::Scalar>::with_capacity(meta.num_challenges);
-
-        let unusable_rows_start = params.n() as usize - (meta.blinding_factors() + 1);
-        let phases = pk.vk.cs.phases().collect::<Vec<_>>();
-        let num_phases = phases.len();
-        // WARNING: this will currently not work if `circuits` has more than 1 circuit
-        // because the original API squeezes the challenges for a phase after running all circuits
-        // once in that phase.
-        if num_phases > 1 {
-            assert_eq!(
-                circuits.len(),
-                1,
-                "New challenge API doesn't work with multiple circuits yet"
-            );
-        }
-        for ((circuit, instances), instance_single) in
-            circuits.iter().zip(instances).zip(instance.iter())
-        {
-            let mut witness: WitnessCollection<Scheme, P, _, E, _, _> = WitnessCollection {
-                params,
-                current_phase: phases[0],
-                advice: vec![domain.empty_lagrange_assigned(); meta.num_advice_columns],
-                instances,
-                challenges: &mut challenges,
-                // The prover will not be allowed to assign values to advice
-                // cells that exist within inactive rows, which include some
-                // number of blinding factors and an extra row for use in the
-                // permutation argument.
-                usable_rows: ..unusable_rows_start,
-                advice_single: AdviceCommitted::<Scheme::Curve, LagrangeCoeff> {
-                    advice_polys: vec![domain.empty_lagrange(); meta.num_advice_columns],
-                    advice_blinds: vec![Blind::default(); meta.num_advice_columns],
-                },
-                instance_single,
-                rng: &mut rng,
-                transcript: &mut transcript,
-                column_indices: column_indices.clone(),
-                challenge_indices: challenge_indices.clone(),
-                unusable_rows_start,
-                _marker: PhantomData,
-            };
-
-            // while loop is for compatibility with circuits that do not use the new `next_phase` API to manage phases
-            // If the circuit uses the new API, then the while loop will only execute once
-            while witness.current_phase.to_u8() < num_phases as u8 {
-                #[cfg(feature = "profile")]
-                let syn_time = start_timer!(|| format!(
-                    "Synthesize time starting from phase {} (synthesize may cross multiple phases)",
-                    witness.current_phase.to_u8()
-                ));
-                // Synthesize the circuit to obtain the witness and other information.
-                ConcreteCircuit::FloorPlanner::synthesize(
-                    &mut witness,
-                    circuit,
-                    config.clone(),
-                    meta.constants.clone(),
-                )
-                .unwrap();
-                #[cfg(feature = "profile")]
-                end_timer!(syn_time);
-                if witness.current_phase.to_u8() < num_phases as u8 {
-                    witness.next_phase();
-                }
-            }
-            advice.push(witness.advice_single);
-        }
-
-        assert_eq!(challenges.len(), meta.num_challenges);
-        let challenges = (0..meta.num_challenges)
-            .map(|index| challenges.remove(&index).unwrap())
-            .collect::<Vec<_>>();
-
-        (advice, challenges)
-    };
+    let (advice, challenges) = witness::run_phase1_synthesis::<Scheme, P, E, R, T, ConcreteCircuit>(
+        params,
+        pk,
+        circuits,
+        instances,
+        &instance,
+        &mut rng,
+        transcript,
+    )?;
     #[cfg(feature = "profile")]
     end_timer!(phase1_time);
 
@@ -507,7 +100,7 @@ where
 /// Restricted to single-circuit, single-phase circuits: multi-phase proving
 /// interleaves challenge squeezes with synthesis and cannot be expressed as a
 /// pre-synthesized advice hand-off.
-pub fn create_proof_raw<
+pub fn create_proof_from_advice<
     'params,
     'a,
     Scheme: CommitmentScheme,
@@ -519,7 +112,7 @@ pub fn create_proof_raw<
     params: &'params Scheme::ParamsProver,
     pk: &ProvingKey<Scheme::Curve>,
     instances: &'a [&'a [Scheme::Scalar]],
-    mut advice: AdviceColumns<Scheme::Scalar>,
+    mut advice: witness::AdviceColumns<Scheme::Scalar>,
     mut rng: R,
     transcript: &'a mut T,
 ) -> Result<(), Error>
@@ -535,21 +128,21 @@ where
     assert_eq!(
         phases.len(),
         1,
-        "create_proof_raw supports single-phase circuits only: multi-phase \
+        "create_proof_from_advice supports single-phase circuits only: multi-phase \
          challenge squeezes interleave with synthesis and cannot go through \
-         create_proof_raw"
+         create_proof_from_advice"
     );
     assert_eq!(
         advice.len(),
         meta.num_advice_columns,
-        "create_proof_raw: advice column count mismatch"
+        "create_proof_from_advice: advice column count mismatch"
     );
     let n = params.n() as usize;
     for col in advice.iter() {
         assert_eq!(
             col.len(),
             n,
-            "create_proof_raw: advice column length must equal params.n()"
+            "create_proof_from_advice: advice column length must equal params.n()"
         );
     }
 
@@ -559,35 +152,7 @@ where
     let domain = &pk.vk.domain;
 
     // Materialize the single instance
-    let instance_single: InstanceSingle<Scheme::Curve> = {
-        let instance_values = instances
-            .iter()
-            .map(|values| {
-                let mut poly = domain.empty_lagrange();
-                assert_eq!(poly.len(), n);
-                if values.len() > (poly.len() - (meta.blinding_factors() + 1)) {
-                    panic!("Error::InstanceTooLarge");
-                }
-                for (poly, value) in poly.iter_mut().zip(values.iter()) {
-                    *poly = *value;
-                }
-                poly
-            })
-            .collect::<Vec<_>>();
-
-        let instance_polys: Vec<_> = instance_values
-            .iter()
-            .map(|poly| {
-                let lagrange_vec = domain.lagrange_from_vec(poly.to_vec());
-                domain.lagrange_to_coeff(lagrange_vec)
-            })
-            .collect();
-
-        InstanceSingle {
-            instance_values,
-            instance_polys,
-        }
-    };
+    let instance_single = build_instance_single::<Scheme>(params, domain, meta, instances);
 
     // Absorb instances into transcript, mirroring `WitnessCollection::next_phase`
     // for phase 0.
@@ -659,10 +224,8 @@ where
     }
 
     // Squeeze phase-0 challenges in constraint-system order.
-    let mut challenge_indices = [(); 3].map(|_| vec![]);
-    for (index, phase) in meta.challenge_phase.iter().enumerate() {
-        challenge_indices[phase.to_u8() as usize].push(index);
-    }
+    let (_column_indices, challenge_indices) =
+        witness::column_and_challenge_indices(meta);
     let mut challenges_map = HashMap::<usize, Scheme::Scalar>::with_capacity(meta.num_challenges);
     for challenge_index in challenge_indices[0].iter() {
         let existing = challenges_map.insert(
@@ -690,192 +253,6 @@ where
         rng,
         transcript,
     )
-}
-
-/// Runs the phase-1 half of proof generation — the same synthesize / commit /
-/// challenge-squeeze steps [`create_proof`] performs internally — and returns
-/// the resulting instance/advice singles and per-phase challenges without
-/// continuing to lookups, permutations, or the multiopen argument.
-///
-/// The transcript is left in the same state [`create_proof`] would produce at
-/// the boundary (vk hashed, instances absorbed, per-phase advice commitments
-/// written, per-phase challenges squeezed). Intended as a diagnostic entry
-/// point — comparing an out-of-band witness generator's advice against
-/// `Circuit::synthesize` — not for producing verifiable proofs.
-///
-/// Mirrors the halo2-gpu fork's `synthesize_witness` shape so downstream crates
-/// can gate on the fork via a single import.
-pub fn synthesize_witness<
-    'params,
-    'a,
-    Scheme: CommitmentScheme,
-    P: Prover<'params, Scheme>,
-    E: EncodedChallenge<Scheme::Curve>,
-    R: RngCore + 'a,
-    T: TranscriptWrite<Scheme::Curve, E>,
-    ConcreteCircuit: Circuit<Scheme::Scalar>,
->(
-    params: &'params Scheme::ParamsProver,
-    pk: &ProvingKey<Scheme::Curve>,
-    circuits: &[ConcreteCircuit],
-    instances: &[&'a [&'a [Scheme::Scalar]]],
-    mut rng: R,
-    mut transcript: &'a mut T,
-) -> Result<
-    (
-        Vec<InstanceSingle<Scheme::Curve>>,
-        Vec<AdviceSingle<Scheme::Curve>>,
-        Vec<Scheme::Scalar>,
-    ),
-    Error,
->
-where
-    Scheme::Scalar: Hash + WithSmallOrderMulGroup<3>,
-    <Scheme as CommitmentScheme>::ParamsProver: Sync,
-{
-    if circuits.len() != instances.len() {
-        return Err(Error::InvalidInstances);
-    }
-
-    for instance in instances.iter() {
-        if instance.len() != pk.vk.cs.num_instance_columns {
-            return Err(Error::InvalidInstances);
-        }
-    }
-
-    // Hash verification key into transcript
-    pk.vk.hash_into(transcript)?;
-
-    let domain = &pk.vk.domain;
-    let mut meta = ConstraintSystem::default();
-    #[cfg(feature = "circuit-params")]
-    let config = ConcreteCircuit::configure_with_params(&mut meta, circuits[0].params());
-    #[cfg(not(feature = "circuit-params"))]
-    let config = ConcreteCircuit::configure(&mut meta);
-    let meta = &pk.vk.cs;
-
-    let instance: Vec<InstanceSingle<Scheme::Curve>> = instances
-        .iter()
-        .map(|instance| -> InstanceSingle<Scheme::Curve> {
-            let instance_values = instance
-                .iter()
-                .map(|values| {
-                    let mut poly = domain.empty_lagrange();
-                    assert_eq!(poly.len(), params.n() as usize);
-                    if values.len() > (poly.len() - (meta.blinding_factors() + 1)) {
-                        panic!("Error::InstanceTooLarge");
-                    }
-                    for (poly, value) in poly.iter_mut().zip(values.iter()) {
-                        *poly = *value;
-                    }
-                    poly
-                })
-                .collect::<Vec<_>>();
-
-            let instance_polys: Vec<_> = instance_values
-                .iter()
-                .map(|poly| {
-                    let lagrange_vec = domain.lagrange_from_vec(poly.to_vec());
-                    domain.lagrange_to_coeff(lagrange_vec)
-                })
-                .collect();
-
-            InstanceSingle {
-                instance_values,
-                instance_polys,
-            }
-        })
-        .collect();
-
-    let mut column_indices = [(); 3].map(|_| vec![]);
-    for (index, phase) in meta.advice_column_phase.iter().enumerate() {
-        column_indices[phase.to_u8() as usize].push(index);
-    }
-    let mut challenge_indices = [(); 3].map(|_| vec![]);
-    for (index, phase) in meta.challenge_phase.iter().enumerate() {
-        challenge_indices[phase.to_u8() as usize].push(index);
-    }
-
-    let mut advice = Vec::with_capacity(instances.len());
-    let mut challenges = HashMap::<usize, Scheme::Scalar>::with_capacity(meta.num_challenges);
-
-    let unusable_rows_start = params.n() as usize - (meta.blinding_factors() + 1);
-    let phases = pk.vk.cs.phases().collect::<Vec<_>>();
-    let num_phases = phases.len();
-    if num_phases > 1 {
-        assert_eq!(
-            circuits.len(),
-            1,
-            "New challenge API doesn't work with multiple circuits yet"
-        );
-    }
-    for ((circuit, circuit_instances), instance_single) in
-        circuits.iter().zip(instances).zip(instance.iter())
-    {
-        let mut witness: WitnessCollection<Scheme, P, _, E, _, _> = WitnessCollection {
-            params,
-            current_phase: phases[0],
-            advice: vec![domain.empty_lagrange_assigned(); meta.num_advice_columns],
-            instances: circuit_instances,
-            challenges: &mut challenges,
-            usable_rows: ..unusable_rows_start,
-            advice_single: AdviceCommitted::<Scheme::Curve, LagrangeCoeff> {
-                advice_polys: vec![domain.empty_lagrange(); meta.num_advice_columns],
-                advice_blinds: vec![Blind::default(); meta.num_advice_columns],
-            },
-            instance_single,
-            rng: &mut rng,
-            transcript: &mut transcript,
-            column_indices: column_indices.clone(),
-            challenge_indices: challenge_indices.clone(),
-            unusable_rows_start,
-            _marker: PhantomData,
-        };
-
-        while witness.current_phase.to_u8() < num_phases as u8 {
-            ConcreteCircuit::FloorPlanner::synthesize(
-                &mut witness,
-                circuit,
-                config.clone(),
-                meta.constants.clone(),
-            )
-            .unwrap();
-            if witness.current_phase.to_u8() < num_phases as u8 {
-                witness.next_phase();
-            }
-        }
-        advice.push(witness.advice_single);
-    }
-
-    assert_eq!(challenges.len(), meta.num_challenges);
-    let challenges = (0..meta.num_challenges)
-        .map(|index| challenges.remove(&index).unwrap())
-        .collect::<Vec<_>>();
-
-    // Convert internal (Lagrange advice + blinds) into the public shape by
-    // dropping blinds and computing the coefficient-basis advice via iFFT.
-    let advice: Vec<AdviceSingle<Scheme::Curve>> = advice
-        .into_iter()
-        .map(|committed| {
-            let AdviceCommitted {
-                advice_polys: advice_values,
-                advice_blinds: _,
-            } = committed;
-            let advice_polys = advice_values
-                .iter()
-                .map(|poly| {
-                    let lagrange_vec = domain.lagrange_from_vec(poly.to_vec());
-                    domain.lagrange_to_coeff(lagrange_vec)
-                })
-                .collect::<Vec<_>>();
-            AdviceSingle {
-                advice_values,
-                advice_polys,
-            }
-        })
-        .collect();
-
-    Ok((instance, advice, challenges))
 }
 
 /// Post-phase-1 half of proof generation: given committed instance/advice
@@ -1207,7 +584,7 @@ where
 fn test_create_proof() {
     use crate::{
         circuit::SimpleFloorPlanner,
-        plonk::{keygen_pk, keygen_vk},
+        plonk::{keygen_pk, keygen_vk, ConstraintSystem},
         poly::kzg::{
             commitment::{KZGCommitmentScheme, ParamsKZG},
             multiopen::ProverSHPLONK,
